+++ grails-app/domain/grails/plugins/DummyDomainClass.groovy
@@ -1,0 +1,7 @@
+package grails.plugins
+
+class DummyDomainClass {
+
+    static constraints = {
+    }
+}
