+++ grails-app/services/grails/plugins/SlugGeneratorService.groovy
@@ -1,4 +1,4 @@
-package grails.slug.generator
+package grails.plugins
 
 import org.grails.core.DefaultGrailsDomainClass
